altoPieza pasa de short a uint8_t y se reduce la altura de las piezas S, T y Z eliminando las filas vacías

+++ escogerPieza.h
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-uint8_t* escogerPieza(short &altoPieza);
+uint8_t* escogerPieza(uint8_t &altoPieza);
 
 #endif
